requirements.txt
# Telegram Bot
aiogram==3.3.0
python-telegram-bot==20.7

# Web App
Flask==3.0.0
Flask-CORS==4.0.0
python-dotenv==1.0.0
gunicorn==21.2.0

# Database
aiosqlite==0.19.0

# Utilities
python-docx==1.0.1
nltk==3.8.1
requests==2.31.0
asyncio==3.4.3
tqdm==4.66.1
pytz==2023.3
python-dateutil==2.8.2

# Development
debugpy==1.8.0
